requirements.txt
amqp==5.2.0
APScheduler==3.9.1.post1
asgiref==3.7.2
async-timeout==5.0.0
beautifulsoup4==4.12.2
billiard==4.2.1
bleach==6.0.0
celery==5.4.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dj-database-url==2.2.0
Django==4.2.4
django-apscheduler==0.6.2
django-bootstrap4==23.2
django-celery-beat==2.7.0
django-ckeditor==6.7.0
django-crispy-forms==1.14.0
django-js-asset==2.1.0
django-reset-migrations==0.4.0
django-timezone-field==7.0
feedparser==6.0.10
gunicorn==22.0.0
idna==3.4
kombu==5.4.2
packaging==24.1
Pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
redis==5.2.0
requests==2.32.3
rq==2.0.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.0.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.8.2
